refactor: rename Base.cpp to EventApiMessage.cpp
Hopefully this fixes the qmake error (Clueless)
