+++ chatterino.pro
@@ -222,9 +222,10 @@ SOURCES += \
     src/providers/seventv/SeventvWebSocket.cpp \
     src/providers/seventv/SeventvEventApiClient.cpp \
     src/providers/seventv/SeventvEventApiManager.cpp \
-    src/providers/seventv/eventapimessages/Base.cpp \
-    src/providers/seventv/eventapimessages/EventApiEmoteUpdate.cpp \
+    src/providers/seventv/SeventvWebSocket.cpp \
+    src/providers/seventv/eventapimessages/EventApiMessage.cpp \
     src/providers/seventv/eventapimessages/EventApiEmoteData.cpp \
+    src/providers/seventv/eventapimessages/EventApiEmoteUpdate.cpp \
     src/providers/twitch/api/Helix.cpp \
     src/providers/twitch/ChannelPointReward.cpp \
     src/providers/twitch/IrcMessageHandler.cpp \
@@ -495,7 +496,8 @@ HEADERS += \
     src/providers/seventv/SeventvEventApiManager.hpp \
     src/providers/seventv/SeventvEventApiMessages.hpp \
     src/providers/seventv/SeventvEventApiWebsocket.hpp \
-    src/providers/seventv/eventapimessages/Base.hpp \
+    src/providers/seventv/SeventvWebSocket.hpp \
+    src/providers/seventv/eventapimessages/EventApiMessage.hpp \
     src/providers/seventv/eventapimessages/EventApiEmoteUpdate.hpp \
     src/providers/seventv/eventapimessages/EventApiEmoteData.hpp \
     src/providers/twitch/api/Helix.hpp \
