+++ src/providers/seventv/SeventvEventApiMessages.hpp
@@ -1,5 +1,5 @@
 #pragma once
 
-#include "providers/seventv/eventapimessages/Base.hpp"
 #include "providers/seventv/eventapimessages/EventApiEmoteData.hpp"
 #include "providers/seventv/eventapimessages/EventApiEmoteUpdate.hpp"
+#include "providers/seventv/eventapimessages/EventApiMessage.hpp"
